OSIS-4527 moved units tests from test_groupelementyear_delete.py to test_detach.py
